Move hostname function from pkg/util/kubernetes to own package
This lets us break an import cycle at the package level (which is
avoided in practice due to build constraints)
