+++ pkg/util/hostname/container_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/config/env"
 	"github.com/DataDog/datadog-agent/pkg/util/docker"
 	"github.com/DataDog/datadog-agent/pkg/util/kubelet"
-	"github.com/DataDog/datadog-agent/pkg/util/kubernetes"
+	k8shostname "github.com/DataDog/datadog-agent/pkg/util/kubernetes/hostname"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -34,7 +34,7 @@ func TestFromContainer(t *testing.T) {
 	defer func() {
 		configIsContainerized = env.IsContainerized
 		configIsFeaturePresent = env.IsFeaturePresent
-		kubernetesGetKubeAPIServerHostname = kubernetes.GetKubeAPIServerHostname
+		kubernetesGetKubeAPIServerHostname = k8shostname.GetKubeAPIServerHostname
 		dockerGetHostname = docker.GetHostname
 		kubeletGetHostname = kubelet.GetHostname
 	}()
@@ -81,7 +81,7 @@ func TestFromContainerInvalidHostname(t *testing.T) {
 	defer func() {
 		configIsContainerized = env.IsContainerized
 		configIsFeaturePresent = env.IsFeaturePresent
-		kubernetesGetKubeAPIServerHostname = kubernetes.GetKubeAPIServerHostname
+		kubernetesGetKubeAPIServerHostname = k8shostname.GetKubeAPIServerHostname
 		dockerGetHostname = docker.GetHostname
 		kubeletGetHostname = kubelet.GetHostname
 	}()
