+++ pkg/util/kubernetes/hostname/hostname_stub.go
@@ -5,7 +5,7 @@
 
 //go:build !kubeapiserver || kubelet
 
-package kubernetes
+package hostname
 
 import (
 	"context"
